# Server/Controllers/StockApi.cs
/*
 * StockQuotes
 *
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 0.1
 * 
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */
using Microsoft.AspNetCore.Mvc;
using Newtonsoft.Json;
using Server.Attributes;
using Server.Models;
using Swashbuckle.AspNetCore.Annotations;
using System;
using System.Collections.Generic;
using System.ComponentModel.DataAnnotations;

namespace Server.Controllers {
	/// <summary>
	/// 
	/// </summary>
	[ApiController]
	public class StockApiController : ControllerBase {
		/// <summary>
		/// 
		/// </summary>
		/// <remarks>Get the information of a stock</remarks>
		/// <param name="id"></param>
		/// <response code="200">Information returned successfully</response>
		/// <response code="404">Resource not found</response>
		[HttpGet]
		[Route("/truemogician/StockQuote/0.1/api/stock/info")]
		[ValidateModelState]
		[SwaggerOperation("GetStockInfo")]
		[SwaggerResponse(statusCode: 200, type: typeof(StockInfo), description: "Information returned successfully")]
		public virtual IActionResult GetStockInfo([FromQuery][Required()][RegularExpression("/\\d{6}\\.[a-zA-Z]{2}/")] string id) {
			//TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
			// return StatusCode(200, default(StockInfo));

			//TODO: Uncomment the next line to return response 404 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
			// return StatusCode(404);
			string exampleJson = null;
			exampleJson = "\"\"";

			var example = exampleJson != null
			? JsonConvert.DeserializeObject<StockInfo>(exampleJson)
			: default(StockInfo);            //TODO: Change the data returned
			return new ObjectResult(example);
		}

		/// <summary>
		/// 
		/// </summary>
		/// <remarks>Get the list of stocks</remarks>
		/// <param name="type"></param>
		/// <param name="date">Only meaningful when quering constituent stock list. Default value is today</param>
		/// <response code="200">List returned successfully</response>
		/// <response code="500">Unknown error on server</response>
		[HttpGet]
		[Route("/truemogician/StockQuote/0.1/api/stock/list")]
		[ValidateModelState]
		[SwaggerOperation("GetStockList")]
		[SwaggerResponse(statusCode: 200, type: typeof(List<StockBasicInfo>), description: "List returned successfully")]
		public virtual IActionResult GetStockList([FromQuery] string type, [FromQuery] DateTime? date) {
			//TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
			// return StatusCode(200, default(List<StockBasicInfo>));

			//TODO: Uncomment the next line to return response 500 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
			// return StatusCode(500);
			string exampleJson = null;
			exampleJson = "[ {\n  \"name\" : \"name\",\n  \"id\" : \"id\"\n}, {\n  \"name\" : \"name\",\n  \"id\" : \"id\"\n} ]";

			var example = exampleJson != null
			? JsonConvert.DeserializeObject<List<StockBasicInfo>>(exampleJson)
			: default(List<StockBasicInfo>);            //TODO: Change the data returned
			return new ObjectResult(example);
		}
	}
}