# Server/Controllers/TokenApi.cs
/*
 * StockQuotes
 *
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 0.1
 * 
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */
using System.ComponentModel.DataAnnotations;
using Microsoft.AspNetCore.Mvc;
using Newtonsoft.Json;
using Server.Attributes;
using Swashbuckle.AspNetCore.Annotations;

namespace Server.Controllers {
	/// <summary>
	/// </summary>
	[ApiController]
	public class TokenApiController : ControllerBase {
		/// <summary>
		/// </summary>
		/// <remarks>Get a new token</remarks>
		/// <response code="201">The new token has been created and returned</response>
		/// <response code="500">Unknown error on server</response>
		[HttpGet]
		[Route("/truemogician/StockQuote/0.1/api/token")]
		[ValidateModelState]
		[SwaggerOperation("GetToken")]
		[SwaggerResponse(201, type: typeof(string), description: "The new token has been created and returned")]
		public virtual IActionResult GetToken() {
			//TODO: Uncomment the next line to return response 201 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
			// return StatusCode(201, default(string));

			//TODO: Uncomment the next line to return response 500 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
			// return StatusCode(500);
			string exampleJson = null;
			exampleJson = "\"\"";

			var example = exampleJson != null
				? JsonConvert.DeserializeObject<string>(exampleJson)
				: default;//TODO: Change the data returned
			return new ObjectResult(example);
		}

		/// <summary>
		/// </summary>
		/// <remarks>Validate token</remarks>
		/// <param name="token"></param>
		/// <response code="200">Validation completed</response>
		/// <response code="500">Unknown error on server</response>
		[HttpGet]
		[Route("/truemogician/StockQuote/0.1/api/token/validation")]
		[ValidateModelState]
		[SwaggerOperation("ValidateToken")]
		[SwaggerResponse(200, type: typeof(bool?), description: "Validation completed")]
		public virtual IActionResult ValidateToken([FromQuery] [Required] string token) {
			//TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
			// return StatusCode(200, default(bool?));

			//TODO: Uncomment the next line to return response 500 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
			// return StatusCode(500);
			string exampleJson = null;
			exampleJson = "true";

			var example = exampleJson != null
				? JsonConvert.DeserializeObject<bool?>(exampleJson)
				: default;//TODO: Change the data returned
			return new ObjectResult(example);
		}
	}
}