# Server/wwwroot/swagger-original.json
{
  "openapi" : "3.0.0",
  "info" : {
    "title" : "StockQuotes",
    "version" : "0.1"
  },
  "servers" : [ {
    "url" : "https://virtserver.swaggerhub.com/truemogician/StockQuote/0.1",
    "description" : "SwaggerHub API Auto Mocking"
  } ],
...
  "paths" : {
    "/api/token" : {
      "get" : {
        "tags" : [ "Token" ],
        "description" : "Get a new token",
        "operationId" : "GetToken",
        "responses" : {
          "201" : {
            "description" : "The new token has been created and returned",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "type" : "string",
                  "description" : "The new token",
                  "x-content-type" : "text/plain"
                }
              }
            }
          },
          "500" : {
            "description" : "Unknown error on server"
          }
        }
      }
    },
    "/api/token/validation" : {
      "get" : {
        "tags" : [ "Token" ],
        "description" : "Validate token",
        "operationId" : "ValidateToken",
        "parameters" : [ {
          "name" : "token",
          "in" : "query",
          "required" : true,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Validation completed",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "type" : "boolean",
                  "description" : "Whether the token is valid",
                  "x-content-type" : "text/plain"
                }
              }
            }
          },
          "500" : {
            "description" : "Unknown error on server"
          }
        }
      }
    },
...
    "/api/stock/list" : {
      "get" : {
        "tags" : [ "Stock" ],
        "description" : "Get the list of stocks",
        "operationId" : "GetStockList",
        "parameters" : [ {
          "name" : "type",
          "in" : "query",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string",
            "default" : "default",
            "enum" : [ "sz50", "hs300", "zz500", "index", "stock", "default" ]
          }
        }, {
          "name" : "date",
          "in" : "query",
          "description" : "Only meaningful when quering constituent stock list. Default value is today",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string",
            "format" : "date"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "List returned successfully",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/StockBasicInfo"
                  },
                  "x-content-type" : "application/json"
                }
              }
            }
          },
          "500" : {
            "description" : "Unknown error on server"
          }
        }
      }
    },
    "/api/stock/info" : {
      "get" : {
        "tags" : [ "Stock" ],
        "description" : "Get the information of a stock",
        "operationId" : "GetStockInfo",
        "parameters" : [ {
          "name" : "id",
          "in" : "query",
          "required" : true,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "pattern" : "\\d{6}\\.[a-zA-Z]{2}",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Information returned successfully",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StockInfo"
                }
              }
            }
          },
          "404" : {
            "description" : "Resource not found"
          }
        }
      }
    },
    "/api/quote/history/minute" : {
      "get" : {
        "tags" : [ "Quote" ],
        "description" : "Get the price summary of a stock",
        "operationId" : "GetMinutelyPrice",
        "parameters" : [ {
          "name" : "id",
          "in" : "query",
          "required" : true,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "pattern" : "\\d{6}\\.[a-zA-Z]{2}",
            "type" : "string"
          }
        }, {
          "name" : "begin-date",
          "in" : "query",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string",
            "format" : "date-time"
          }
        }, {
          "name" : "end-date",
          "in" : "query",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string",
            "format" : "date-time"
          }
        }, {
          "name" : "frequency",
          "in" : "query",
          "description" : "Frequency of the quotes",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 60,
            "enum" : [ 5, 15, 30, 60 ]
          }
        }, {
          "name" : "rehabilitation",
          "in" : "query",
          "description" : "Rehabilitation status, default value is \"none\"",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string",
            "default" : "none",
            "enum" : [ "pre", "post", "none" ]
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Price summary returned successfully",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/MinutelyPrice"
                  },
                  "x-content-type" : "application/json"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid parameters or payload"
          },
          "500" : {
            "description" : "Unknown error on server"
          }
        }
      }
    },
    "/api/quote/history/day" : {
      "get" : {
        "tags" : [ "Quote" ],
        "description" : "Get the price summary of a stock",
        "operationId" : "GetDailyPrice",
        "parameters" : [ {
          "name" : "id",
          "in" : "query",
          "required" : true,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "pattern" : "\\d{6}\\.[a-zA-Z]{2}",
            "type" : "string"
          }
        }, {
          "name" : "begin-date",
          "in" : "query",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string",
            "format" : "date-time"
          }
        }, {
          "name" : "end-date",
          "in" : "query",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string",
            "format" : "date-time"
          }
        }, {
          "name" : "rehabilitation",
          "in" : "query",
          "description" : "Rehabilitation status, default value is \"none\"",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string",
            "default" : "none",
            "enum" : [ "pre", "post", "none" ]
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Price summary returned successfully",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/DailyPrice"
                  },
                  "x-content-type" : "application/json"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid parameters or payload"
          },
          "500" : {
            "description" : "Unknown error on server"
          }
        }
      }
    },
    "/api/quote/history/week" : {
      "get" : {
        "tags" : [ "Quote" ],
        "description" : "Get the price summary of a stock",
        "operationId" : "GetWeeklyPrice",
        "parameters" : [ {
          "name" : "id",
          "in" : "query",
          "required" : true,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "pattern" : "\\d{6}\\.[a-zA-Z]{2}",
            "type" : "string"
          }
        }, {
          "name" : "begin-date",
          "in" : "query",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string",
            "format" : "date-time"
          }
        }, {
          "name" : "end-date",
          "in" : "query",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string",
            "format" : "date-time"
          }
        }, {
          "name" : "frequency",
          "in" : "query",
          "description" : "Frequency of the quotes",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string",
            "default" : "week",
            "enum" : [ "week", "month" ]
          }
        }, {
          "name" : "rehabilitation",
          "in" : "query",
          "description" : "Rehabilitation status, default value is \"none\"",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string",
            "default" : "none",
            "enum" : [ "pre", "post", "none" ]
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Price summary returned successfully",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/WeeklyPrice"
                  },
                  "x-content-type" : "application/json"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid parameters or payload"
          },
          "500" : {
            "description" : "Unknown error on server"
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "StockBasicInfo" : {
        "required" : [ "id", "name" ],
        "type" : "object",
        "properties" : {
          "id" : {
            "pattern" : "[a-zA-Z]{2}\\.\\d{6}",
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          }
        },
        "example" : {
          "name" : "name",
          "id" : "id"
        }
      },
      "StockInfo" : {
        "allOf" : [ {
          "$ref" : "#/components/schemas/StockBasicInfo"
        }, {
          "required" : [ "listedDate", "type" ],
          "type" : "object",
          "properties" : {
            "id" : {
              "pattern" : "[a-zA-Z]{2}\\.\\d{6}",
              "type" : "string"
            },
            "name" : {
              "type" : "string"
            },
            "type" : {
              "type" : "string",
              "enum" : [ "stock", "index", "other" ]
            },
            "industry" : {
              "type" : "string"
            },
            "classification" : {
              "type" : "string"
            },
            "listedDate" : {
              "type" : "string",
              "format" : "date"
            },
            "delistedDate" : {
              "type" : "string",
              "format" : "date"
            }
          }
        } ]
      },
      "PriceBase" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "pattern" : "[a-zA-Z]{2}\\.\\d{6}",
            "type" : "string"
          },
          "preClosing" : {
            "type" : "integer",
            "format" : "int32"
          },
          "opening" : {
            "type" : "integer",
            "format" : "int32"
          },
          "closing" : {
            "type" : "integer",
            "format" : "int32"
          },
          "highest" : {
            "type" : "integer",
            "format" : "int32"
          },
          "lowest" : {
            "type" : "integer",
            "format" : "int32"
          },
          "volume" : {
            "type" : "integer",
            "format" : "int32"
          },
          "turnover" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
...
      "DailyPrice" : {
        "allOf" : [ {
          "$ref" : "#/components/schemas/PriceBase"
        }, {
          "type" : "object",
          "properties" : {
            "date" : {
              "type" : "string",
              "format" : "date"
            },
            "rehabilitation" : {
              "type" : "string",
              "enum" : [ "pre", "post", "none" ]
            },
            "turnoverRate" : {
              "pattern" : "\\d{6}",
              "type" : "string"
            },
            "per" : {
              "pattern" : "\\d{6}",
              "type" : "string",
              "description" : "price/earnings ratio"
            },
            "psr" : {
              "pattern" : "\\d{6}",
              "type" : "string",
              "description" : "price to sales ratio"
            },
            "pcfr" : {
              "pattern" : "\\d{6}",
              "type" : "string",
              "description" : "price cash flow ratio"
            },
            "pbr" : {
              "pattern" : "\\d{6}",
              "type" : "string",
              "description" : "price to book ratio"
            },
            "stopped" : {
              "type" : "boolean"
            },
            "specialTreatment" : {
              "type" : "boolean"
            }
          }
        } ]
      },
      "MinutelyPrice" : {
        "allOf" : [ {
          "$ref" : "#/components/schemas/PriceBase"
        }, {
          "type" : "object",
          "properties" : {
            "time" : {
              "type" : "string",
              "format" : "date-time"
            },
            "rehabilitation" : {
              "type" : "string",
              "enum" : [ "pre", "post", "none" ]
            }
          }
        } ]
      },
      "WeeklyPrice" : {
        "allOf" : [ {
          "$ref" : "#/components/schemas/PriceBase"
        }, {
          "type" : "object",
          "properties" : {
            "date" : {
              "type" : "string",
              "format" : "date"
            },
            "rehabilitation" : {
              "type" : "string",
              "enum" : [ "pre", "post", "none" ]
            },
            "turnoverRate" : {
              "type" : "integer",
              "format" : "int32"
            }
          }
        } ]
      },
...
    },
...
    "securitySchemes" : {
      "token" : {
        "type" : "apiKey",
        "name" : "token",
        "in" : "header"
      }
    }
  }
}